post_add, post_edit

+++ django/blog/views.py
@@ -1,91 +1,133 @@
 from django.shortcuts import render, redirect
-from django.http import HttpResponse
+# from django.contrib import messages
+
 from .models import Post
 
 
-# Create your views here.
-
 def post_list(request):
     # 1. 브라우저에서 요청
-    # 2. 장고 runserver로 실행중인 서버에 도착
-    # 3. runserver는 요청을 장고코드로 전달
-    # 4. 장고 코드중 config.urls묘듈이 해당요청을 받음
-    # 5. config.urls모듈은 ''(admin/를 제외한 모든요청)을 blog.urls모듈로 전달
-    # 6. blog.urls모듈은 받은 요청의 url과 일치하는 패턴이 있는지 검사
+    # 2. 요청이 runserver로 실행중인 서버에 도착
+    # 3. runserver는 요청을 Django code로 전달
+    # 4. Django code중 config.urls모듈이 해당 요청을 받음
+    # 5. config.urls모듈은 ''(admin/를 제외한 모든 요청)을 blog.urls모듈로 전달
+    # 6. blog.urls모듈은 받은 요청의 URL과 일치하는 패턴이 있는지 검사
     # 7. 있다면 일치하는 패턴과 연결된 함수(view)를 실행
-    #   7.1 setting모듈의 Templates 속성 내의 목록에서 blog/post_list.html파일의 내용을 가져옴
-    #   7.2 가져온 내용을 처리 (렌더렝, render())하여 리턴
-    # 8. 함수의 실행결과(리턴값)을 브라우저로 다시전달
-    # Http프로토콜로 텍스트 데이터 응답을 반환
+    #   7-1. settings모듈의 TEMPLATES속성 내의 DIRS목록에서
+    #        blog/post_list.html파일의 내용을 가져옴
+    #   7-2. 가져온 내용을 적절히 처리(렌더링, render()함수)하여 리턴
+    # 8. 함수의 실행 결과(리턴값)를 브라우저로 다시 전달
 
+    # HTTP프로토콜로 텍스트 데이터 응답을 반환
+    # return HttpResponse('<html><body><h1>Post list</h1><p>Post목록을 보여줄 예정입니다</p></body></html>')
+
+    # 구글검색: django model order recently created in view
+    # order_by
     posts = Post.objects.all()
-    # render()함수에 전달할 dict 객체 생성
-
+    # render()함수에 전달할 dict객체 생성
     context = {
         'posts': posts,
     }
-
-    # return HttpResponse('Post List')
-    # return HttpResponse('<html><bodㅇy><h1>Post List</h1><p>post 목록을 보여줄 예정입니다.</p></body></html>')
     return render(
         request=request,
         template_name='blog/post_list.html',
         context=context,
     )
-    # post_detail('localhost:8000/detail')
+    # 위 return코드와 같음
+    # return render(request, 'blog/post_list.html', context)
 
 
 def post_detail(request, pk):
     context = {
         'post': Post.objects.get(pk=pk),
     }
-    return render(request, 'blog/post_detail.html', context)
-    '''
-    localhost:8000/detail/로 온 요청을 'blog/post_detail.html'을 render한 결과를 리턴 
-    :return: 
-    '''
+    return render(
+        request,
+        'blog/post_detail.html',
+        context
+    )
+
+
+def post_edit(request, pk):
+    """
+    1. pk에 해당하는 Post인스턴스를
+        context라는 dict에 'post'키에 할당
+        위에서 생성한 dict는 render의 context에 전달
+        사용하는 템플릿은 'blog/post_add.html'을 재사용
+            HTML새로 만들지 말고 있던 html을 그냥 할당
+    2. url은 /<pk>/edit/ <- 에 매칭되도록 urls.py작성
+    3. 이 위치로 올 수 있는 a요소를 post_detail.html에 작성 (form아님)
+    - request.method가 POST일 때는 request.POST에 있는 데이터를 이용해서
+      pk에 해당하는 Post인스턴스의 값을 수정, 이후 post-detail로 redirect
+        값을 수정하는 코드
+            post = Post.objects.get(pk=pk)
+            post.title = <새 문자열>
+            post.content = <새 문자열>
+            post.save()  <- DB에 업데이트 됨
+    - request.method가 GET일 때는 현재 아래에 있는 로직을 실행
+    :param request:
+    :param pk:
+    :return:
+    """
+    # 현재 URL (pk가 3일경우 /3/edit/)에 전달된 pk에 해당하는 Post인스턴스를 post변수에 할당
+    post = Post.objects.get(pk=pk)
+    # 만약 POST메서드 요청일 경우
+    if request.method == 'POST':
+        # post의 제목/내용을 전송받은 값으로 수정 후
+        title = request.POST['title']
+        content = request.POST['content']
+        post.title = title
+        post.content = content
+        # DB에 저장
+        post.save()
+        # 이후 상세화면으로 이동
+        return redirect('post-detail', pk=post.pk)
+    # GET메서드 요청일 경우
+    context = {
+        'post': post,
+    }
+    # 수정할 수 있는 페이지를 보여줌
+    return render(request, 'blog/post_edit.html', context)
 
 
 def post_add(request):
-    # 1. localhost:8000/add로 접근시
-    # 이 뷰가 실행되어서 pOST ADD PAGE 라는 문구를 보여주도록 urls 작
-    # 2.  HttpResponse가 아니라 blog/post_add.html을 출력
-    # post_add.html은 base.html을 확장 title(h2)부분에 'post add' 출력
-
+    # localhost:8000/add로 접근시
+    # 이 뷰가 실행되어서 Post add page라는 문구를 보여주도록 urls작성
+    # HttpResponse가 아니라 blog/post_add.html을 출력
+    # post_add.html은 base.html을 확장, title(h2)부분에 'Post add'라고 출력
     if request.method == 'POST':
-        # 요청의 method가 post일 때
+        # 요청의 method가 POST일 때
+        # HttpResponse로 POST요청에 담겨온
+        # title과 content를 합친 문자열 데이터를 보여줌
+
         title = request.POST['title']
         content = request.POST['content']
-        # return HttpResponse(f'{title}:{content}')
-        # ORM을 사용해서 title과 content 해당하는 post생성
 
-        post = Post.objects.create(
-            author=request.user,
-            title=title,
-            content=content,
-        )
-        return redirect('post-detail', pk=post.pk)
+        #만약 title 이나 content가 비어있으면
+        # 다시 글 작성화면 으로 이동
+        # 이동시키지 말고 오류메세지를 출력
+        # if title is None or content is None :
 
-        #redirect가 아니라 render를 쓰는경우
-        # context ={
-        #     'post' : post,
-        # }
-        # return render(request ,'blog/post_detail.html',context)
 
-        #return HttpResponse(f'{post.pk}{post.title}{post.content}')
+        if not (title and content):
+            content['form_error'] = '제목과 내용을 입력해주세요'
+        else :
+        # ORM을 사용해서 title과 content에 해당하는 Post생성
+            post = Post.objects.create(
+                author=request.user,
+                title=title,
+                content=content,
+            )
+        # post-detail이라는 URL name을 가진 뷰로
+        # 리디렉션 요청을 보냄
+        # 이 때, post-detail URL name으로 특정 URL을 만드려면
+        # pk값이 필요하므로 키워드 인수로 해당 값을 넘겨준다
+            return redirect('post-detail', pk=post.pk)
+    # 요청의 method가 GET일 때
 
-    else:
-        # 요청의 method가 get일때
-        pass
+        return render(request, 'blog/post_add.html', context)
 
-        return render(request, 'blog/post_add.html')
-    # return HttpResponse('Post add page')
-#
-# def post_delete(request, pk):
-#     if Post.request =='POST' :
-#         post = Post.obeject.get(pk=pk)
-#         post.delete()
-#         return redirect('post-list')
+
+
 
 def post_delete(request, pk):
     """
@@ -104,9 +146,13 @@ def post_delete(request, pk):
         csrf_token사용!
         action의 위치가 요청을 보낼 URL임
     """
-    # pk에 해당하는 Post를 삭제
-    if request.method =='POST' :
+    if request.method == 'POST':
+        # pk에 해당하는 Post를 삭제
         post = Post.objects.get(pk=pk)
-        post.delete()
-        # 이후 post-list라는 URL name을 갖는 view로 redirect
-        return redirect('post-list')
+        # 삭제 요청한 user와 post의 author가 같을때만 해당 post를 삭제
+        if request.user == post.author:
+            post.delete()
+            # 이후 post-list라는 URL name을 갖는 view로 redirect
+            return redirect('post-list')
+        # 요청한 유저가 다르면 다시 글 상세화면으로 돌아옴
+        return redirect('post-detail', pk=post.pk)
